// src/api/generated/models/staffDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { StaffDtoStatus } from './staffDtoStatus';
import type { StaffScopeDto } from './staffScopeDto';

export interface StaffDto {
  id?: string;
  username?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  phone?: string;
  status?: StaffDtoStatus;
  createdAt?: string;
  scopes?: StaffScopeDto[];
}